refactored facilities code
