+++ frontend/src/routes/FacilityBooking/CreateBooking/index.tsx
@@ -1,14 +1,17 @@
 import React, { useEffect, useState } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
 import { useHistory, useParams } from 'react-router-dom'
 import { useForm } from 'react-hook-form'
-import { useDispatch, useSelector } from 'react-redux'
 import styled from 'styled-components'
 
 import { AutoComplete } from 'antd'
-import TopNavBar from '../../../components/Mobile/TopNavBar'
 import 'antd-mobile/dist/antd-mobile.css'
 import 'antd/dist/antd.css'
-import { RootState } from '../../../store/types'
+
+import { PATHS } from '../../Routes'
+import { unixToFullDate } from '../../../common/unixToFullDate'
+import { get24Hourtime } from '../../../common/get24HourTime'
+
 import {
   fetchAllCCAs,
   getFacilityList,
@@ -21,16 +24,16 @@ import {
   setSelectedEndTime,
   setBookingStatus,
 } from '../../../store/facilityBooking/action'
+import { RootState } from '../../../store/types'
+import { BookingStatus } from '../../../store/facilityBooking/types'
+
+import TopNavBar from '../../../components/Mobile/TopNavBar'
 import LoadingSpin from '../../../components/LoadingSpin'
-import { PATHS } from '../../Routes'
 import InputField from '../../../components/Mobile/InputField'
 // TODO, change the way how InputField works?
 import { Switch } from '../../../components/Switch'
-import { BookingStatus } from '../../../store/facilityBooking/types'
-import { unixToFullDate } from '../../../common/unixToFullDate'
 import SelectableField from '../../../components/SelectableField'
 import ButtonComponent from '../../../components/Button'
-import { get24Hourtime } from '../../../common/get24HourTime'
 import ConflictBookingModal from '../ViewConflicts/ConflictBookingModal'
 
 const Background = styled.div`
