+++ frontend/src/routes/FacilityBooking/EditBooking/index.tsx
@@ -2,13 +2,14 @@ import React, { useEffect, useState } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
 import { useForm } from 'react-hook-form'
 import { useDispatch, useSelector } from 'react-redux'
+
 import styled from 'styled-components'
-import axios from 'axios'
 import { AutoComplete } from 'antd'
-import TopNavBar from '../../../components/Mobile/TopNavBar'
 import 'antd-mobile/dist/antd-mobile.css'
 import 'antd/dist/antd.css'
-import { RootState } from '../../../store/types'
+
+import { PATHS } from '../../Routes'
+import ConflictBookingModal from '../ViewConflicts/ConflictBookingModal'
 import {
   fetchAllCCAs,
   getFacilityList,
@@ -21,8 +22,10 @@ import {
   resetNewBooking,
   fetchSelectedFacility,
 } from '../../../store/facilityBooking/action'
+import { RootState } from '../../../store/types'
+
 import LoadingSpin from '../../../components/LoadingSpin'
-import { PATHS } from '../../Routes'
+import TopNavBar from '../../../components/Mobile/TopNavBar'
 import InputField from '../../../components/Mobile/InputField'
 // TODO, change the way how InputField works?
 import { Switch } from '../../../components/Switch'
@@ -31,7 +34,6 @@ import { unixToFullDate } from '../../../common/unixToFullDate'
 import SelectableField from '../../../components/SelectableField'
 import ButtonComponent from '../../../components/Button'
 import { get24Hourtime } from '../../../common/get24HourTime'
-import ConflictBookingModal from '../ViewConflicts/ConflictBookingModal'
 
 const Background = styled.div`
   background-color: #fff;
@@ -168,92 +170,97 @@ export default function EditBooking() {
     if (bookingStatus === BookingStatus.SUCCESS) {
       history.replace(PATHS.FACILITY_BOOKING_MAIN)
       history.push(`${PATHS.VIEW_FACILITY}/${params.facilityId}`)
-    }
-    if (bookingStatus === BookingStatus.CONFLICT) {
+    } else if (bookingStatus === BookingStatus.CONFLICT) {
       setmodalIsOpen(true)
     }
   }, [bookingStatus])
 
   return (
     <Background>
-      {!isLoading && !selectedBooking ? (
+      {isLoading ? (
         <LoadingSpin />
       ) : (
-        <>
-          <TopNavBar title={`Edit Booking for ${selectedBooking?.facilityName}`} />
-          <Form onSubmit={onSubmit}>
-            <InputField
-              name={'eventName'}
-              title="Event Name"
-              placeholder="Event Name"
-              required
-              register={register}
-              defaultValue={selectedBooking?.eventName}
-              setValue={setValue}
-              errors={errors.eventName}
-            />
-            <SelectableField
-              title="Start"
-              value={
-                bookingStartTime == 0 ? '' : unixToFullDate(bookingStartTime) + ' at ' + get24Hourtime(bookingStartTime)
-              }
-              isCompulsory={true}
-              onClick={() => dispatch(setBookingStartTime(1644641028))}
-            ></SelectableField>
-            <SelectableField
-              title="End"
-              value={bookingEndTime == 0 ? '' : unixToFullDate(bookingEndTime) + ' at ' + get24Hourtime(bookingEndTime)}
-              isCompulsory={true}
-              onClick={() => dispatch(setBookingEndTime(1644648228))}
-            ></SelectableField>
-            <Container>
-              <StyledTitle>CCA</StyledTitle>
-              <CCAInput
-                options={ccaList.concat({ ccaID: 0, ccaName: 'Personal', category: 'Personal' }).map((cca) => ({
-                  value: cca.ccaName,
-                }))}
-                value={selectedBooking?.ccaName}
-                placeholder="CCAs"
-                onChange={(value) => setCcaName(value)}
-                filterOption={(inputValue, option) =>
-                  option?.value.toUpperCase().indexOf(inputValue.toUpperCase()) !== -1
-                }
-                notFoundContent="No Matching CCAs"
-                allowClear
+        selectedBooking && (
+          <>
+            <TopNavBar title={`Edit Booking for ${selectedBooking.facilityName}`} />
+            <Form onSubmit={onSubmit}>
+              <InputField
+                name={'eventName'}
+                title="Event Name"
+                placeholder="Event Name"
+                required
+                register={register}
+                defaultValue={selectedBooking.eventName}
+                setValue={setValue}
+                errors={errors.eventName}
               />
-            </Container>
-            <InputField
-              name={'description'}
-              title="Description"
-              placeholder="Tell us what your booking is for!"
-              textArea
-              defaultValue={selectedBooking?.description}
-              register={register}
-              setValue={setValue}
-              errors={errors.description}
-            />
-            <WeeklyRecurrenceRow>
-              <StyledTitle>Weekly Recurrence</StyledTitle>
-              <Switch isOn={isWeeklyOn} handleToggle={() => setIsWeeklyOn(!isWeeklyOn)} switchSize={50} />
-            </WeeklyRecurrenceRow>
-            {isWeeklyOn && (
               <SelectableField
-                title="End Date"
-                value={bookingEndDate == 0 ? '' : unixToFullDate(bookingEndDate)}
+                title="Start"
+                value={
+                  bookingStartTime == 0
+                    ? ''
+                    : unixToFullDate(bookingStartTime) + ' at ' + get24Hourtime(bookingStartTime)
+                }
                 isCompulsory={true}
-                onClick={() => dispatch(setBookingEndDate(1644648228))}
-              ></SelectableField>
-            )}
-            <ConflictBookingModal modalOpen={modalIsOpen} setModalOpen={setmodalIsOpen} />
-            <ButtonComponent
-              state={'primary'}
-              text="Submit"
-              type="submit"
-              disabled={!formIsValid()}
-              onClick={() => console.log('submitted')}
-            ></ButtonComponent>
-          </Form>
-        </>
+                onClick={() => dispatch(setBookingStartTime(1644641028))}
+              />
+              <SelectableField
+                title="End"
+                value={
+                  bookingEndTime == 0 ? '' : unixToFullDate(bookingEndTime) + ' at ' + get24Hourtime(bookingEndTime)
+                }
+                isCompulsory={true}
+                onClick={() => dispatch(setBookingEndTime(1644648228))}
+              />
+              <Container>
+                <StyledTitle>CCA</StyledTitle>
+                <CCAInput
+                  options={ccaList.concat({ ccaID: 0, ccaName: 'Personal', category: 'Personal' }).map((cca) => ({
+                    value: cca.ccaName,
+                  }))}
+                  value={selectedBooking.ccaName}
+                  placeholder="CCAs"
+                  onChange={(value) => setCcaName(value)}
+                  filterOption={(inputValue, option) =>
+                    option?.value.toUpperCase().indexOf(inputValue.toUpperCase()) !== -1
+                  }
+                  notFoundContent="No Matching CCAs"
+                  allowClear
+                />
+              </Container>
+              <InputField
+                name={'description'}
+                title="Description"
+                placeholder="Tell us what your booking is for!"
+                textArea
+                defaultValue={selectedBooking.description}
+                register={register}
+                setValue={setValue}
+                errors={errors.description}
+              />
+              <WeeklyRecurrenceRow>
+                <StyledTitle>Weekly Recurrence</StyledTitle>
+                <Switch isOn={isWeeklyOn} handleToggle={() => setIsWeeklyOn(!isWeeklyOn)} switchSize={50} />
+              </WeeklyRecurrenceRow>
+              {isWeeklyOn && (
+                <SelectableField
+                  title="End Date"
+                  value={bookingEndDate == 0 ? '' : unixToFullDate(bookingEndDate)}
+                  isCompulsory={true}
+                  onClick={() => dispatch(setBookingEndDate(1644648228))}
+                />
+              )}
+              <ConflictBookingModal modalOpen={modalIsOpen} setModalOpen={setmodalIsOpen} />
+              <ButtonComponent
+                state={'primary'}
+                text="Submit"
+                type="submit"
+                disabled={!formIsValid()}
+                onClick={() => console.log('submitted')}
+              />
+            </Form>
+          </>
+        )
       )}
     </Background>
   )
